Advanced State Management System Context + useReducer

+++ 11-world-wise/src/contexts/CitiesContext.jsx
@@ -1,32 +1,76 @@
 /** @format */
 
-import { createContext, useContext, useEffect, useState } from "react";
+import { createContext, useContext, useEffect, useReducer } from "react";
 const BASE_URL = "http://localhost:9000";
 
 const CitiesContext = createContext();
 
+const initialState = {
+  cities: [],
+  isLoading: false,
+  currentCity: {},
+};
+
+function reducer(state, action) {
+  switch (action.type) {
+    case "loading":
+      return {
+        ...state,
+        isLoading: true,
+      };
+    case "cities/loaded":
+      return {
+        ...state,
+        isLoading: false,
+        cities: action.payload,
+      };
+    case "city/loaded":
+      return { ...state, isLoading: false, currentCity: action.payload };
+    case "city/created":
+      return {
+        ...state,
+        isLoading: false,
+        cities: [...state.cities, action.payload],
+        currentCity: action.payload,
+      };
+    case "city/deleted":
+      return {
+        ...state,
+        isLoading: false,
+        cities: state.cities.filter((city) => city.id !== action.payload),
+        currentCity: {},
+      };
+    case "rejected":
+      return {
+        ...state,
+        isLoading: false,
+        error: action.payload,
+      };
+    default:
+      throw new Error("Unknown action type");
+  }
+}
 function CitiesProvider({ children }) {
-  const [cities, setCities] = useState([]);
-  const [isLoading, setIsLoading] = useState(false);
-  const [currentCity, setCurrentCity] = useState({});
-
-
+  const [{ cities, isLoading, currentCity, error }, dispatch] = useReducer(
+    reducer,
+    initialState
+  );
   // fetch city data
   useEffect(function () {
     async function fetchCities() {
       try {
-        setIsLoading(true);
+        dispatch({ type: "loading" });
         const res = await fetch(`${BASE_URL}/cities`);
         const data = await res.json();
 
         // set the city data inside the state array
-        setCities(data);
+        dispatch({ type: "cities/loaded", payload: data });
       } catch {
         // Error handling
-        alert("There was an error loading data...");
-      } finally {
-        // setIsLoading is off on all the effect except one place
-        setIsLoading(false);
+        dispatch({
+          type: "rejected",
+          payload: "There was an error loading cities...",
+        });
       }
     }
     // calling the function
@@ -35,24 +79,25 @@ function CitiesProvider({ children }) {
   // get the city data based on id
   async function getCity(id) {
     try {
-      setIsLoading(true);
+      dispatch({ type: "loading" });
       // Different API
       // the data in this case is the city object with the id passed to the parameters
       const res = await fetch(`${BASE_URL}/cities/${id}`);
       const data = await res.json();
       // set the data into another object which is a state
-      setCurrentCity(data);
+      dispatch({ type: "city/loaded", payload: data });
     } catch {
       // error handling
-      alert("There was an error loading data...");
-    } finally {
-      setIsLoading(false);
+      dispatch({
+        type: "rejected",
+        payload: "There was an error loading the city...",
+      });
     }
   }
 
   async function createCity(newCity) {
     try {
-      setIsLoading(true);
+      dispatch({ type: "loading" });
       // Different API
       // POST the stringified new city to the API
       const res = await fetch(`${BASE_URL}/cities`, {
@@ -64,39 +109,44 @@ function CitiesProvider({ children }) {
       });
       const data = await res.json();
       // add the data to the cities state and create an array of old states and new state values
-      setCities((cities) => [...cities, data]);
+      dispatch({ type: "city/created", payload: data });
     } catch {
       // Error handling
-      alert("There was an error creating city ");
-    } finally {
-      setIsLoading(false);
+      dispatch({
+        type: "rejected",
+        payload: "There was an error creating city ",
+      });
     }
   }
   async function deleteCity(id) {
+    if (Number(id) === currentCity.id) return;
+
     try {
-      setIsLoading(true);
+      dispatch({ type: "loading" });
       // Different API
       // DELETE the city data from the API
       const res = await fetch(`${BASE_URL}/cities/${id}`, {
         method: "DELETE",
       });
       // filter the cities array and remove the city
-      setCities((cities) => cities.filter((city) => city.id !== id));
+      dispatch({ type: "city/deleted", payload: id });
     } catch {
       // Error handling
-      alert("There was an error deleting city");
-    } finally {
-      setIsLoading(false);
+      dispatch({
+        type: "rejected",
+        payload: "There was an error deleting the city",
+      });
     }
   }
 
   return (
-    // returning a context provider with the values 
+    // returning a context provider with the values
     <CitiesContext.Provider
       value={{
         cities,
         isLoading,
         currentCity,
+        error,
         getCity,
         createCity,
         deleteCity,
@@ -113,4 +163,3 @@ function useCities() {
   return context;
 }
 export { CitiesProvider, useCities };
-
